% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeTrackVIz2_main.R
\name{plot_bed}
\alias{plot_bed}
\title{bed region plotter}
\usage{
plot_bed(bed, name_bed, color_bed, size_bed, mychr, start_loc, end_loc,
  breaks_wanted = 4, y_title = "SE", show_partial_overlap = T,
  same_y_level = T, axis_line_size = 1, marginvec_mm = c(0, 0, 0, 0),
  show_labels = F, print_plot = F)
}
\arguments{
\item{bed}{bedpe file for DNA loops or interaction data}

\item{name_bed}{name of sample}

\item{color_bed}{color of track}

\item{size_bed}{size of bed (height)}

\item{mychr}{chromosome}

\item{start_loc}{start location on chromomse}

\item{end_loc}{end location on chromosome}

\item{breaks_wanted}{number of breaks on x-axis}

\item{y_title}{title for y-axis}

\item{show_partial_overlap}{include loops that do not entirely overlap the coordinates}

\item{same_y_level}{plot all bed regions on same height (y-axis)}

\item{axis_line_size}{line size of x-axis}

\item{marginvec_mm}{margin around plot, default = c(0,0,0,0)}

\item{show_labels}{boolean: show x-axis lables}

\item{print_plot}{boolean: print individual plot}
}
\value{
ggplot object
}
\description{
This function plots data from bed file
}
\examples{
    plot_bed(bed)


}
\keyword{bedpe}
\keyword{loops,}
