% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeTrackVIz2_main.R
\name{plot_bedgraph}
\alias{plot_bedgraph}
\title{DNA bedgraph plotter}
\usage{
plot_bedgraph(bdg, name_bdg, color_bdg, mychr, start_loc, end_loc,
  binsize = 100, breaks_wanted = 4, y_title = "RPM/bp",
  chrom_col = "chr", start_col = "start", end_col = "end",
  counts_col = "RPKM", forced_y_min = NULL, forced_y_max = NULL,
  axis_line_size = 0.2, marginvec_mm = c(0, 0, 0, 0),
  use_barplot = F, alpha_fill = 1, reverse_strand = FALSE,
  show_labels = F, print_plot = F)
}
\arguments{
\item{bdg}{bedgraph file for DNA coverage data}

\item{name_bdg}{name of sample}

\item{color_bdg}{color of track}

\item{mychr}{chromosome}

\item{start_loc}{start location on chromomse}

\item{end_loc}{end location on chromosome}

\item{binsize}{size of bin}

\item{breaks_wanted}{number of breaks on x-axis}

\item{y_title}{title for y-axis}

\item{chrom_col}{column name for chromosomes}

\item{start_col}{column name for start}

\item{end_col}{column name for end}

\item{counts_col}{column name with counts}

\item{forced_y_min}{hard minimum y-axis value}

\item{forced_y_max}{hard maximum y-axis value}

\item{axis_line_size}{line size of x-axis}

\item{marginvec_mm}{margin around plot, default = c(0,0,0,0)}

\item{use_barplot}{plot track as barplot, default = F}

\item{alpha_fill}{alpha for lineplot, default = 1}

\item{reverse_strand}{boolean: flip strand x-axis direction}

\item{show_labels}{boolean: show x-axis lables}

\item{print_plot}{boolean: print individual plot}
}
\value{
ggplot object
}
\description{
This function plots DNA bedgraph data
}
\examples{
    plot_bedgraph(bdg)


}
\keyword{DNA,}
\keyword{bedgraph}
\keyword{coverage,}
