% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeTrackVIz2_main.R
\name{plot_coord}
\alias{plot_coord}
\title{genomic coordinate plotter}
\usage{
plot_coord(mychr, start_loc, end_loc, breaks_wanted = 4,
  label_size = 6, axis_line_size = 0.5, marginvec_mm = c(0, 0, 0, 0),
  reverse_strand = FALSE, show_labels = T, print_plot = F)
}
\arguments{
\item{mychr}{chromosome}

\item{start_loc}{start location on chromomse}

\item{end_loc}{end location on chromosome}

\item{breaks_wanted}{number of breaks on x-axis}

\item{label_size}{label size for x-axis text, default is 6}

\item{axis_line_size}{line size of x-axis}

\item{marginvec_mm}{margin around plot, default = c(0,0,0,0)}

\item{show_labels}{boolean: show x-axis lables}

\item{print_plot}{boolean: print individual plot}
}
\value{
ggplot object
}
\description{
This function plots genomic coordinates
}
\examples{
    plot_coord(mychr, ...)


}
\keyword{bedpe}
\keyword{loops,}
