% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeTrackVIz2_main.R
\name{plot_genome}
\alias{plot_genome}
\title{genome model plotter}
\usage{
plot_genome(transcript, exon, five_UTR, three_UTR, mychr = "chr14",
  start_loc = 61100000, end_loc = 61195000, breaks_wanted = 4,
  show_partial_overlap = T, marginvec_mm = c(0, 0, 0, 0),
  y_coord_gene_dist = 0.5, y_top_space = 2, y_genename_space = 1,
  genename_size = 4, exon_size = 0.5, UTR_size = 0.5,
  reverse_strand = FALSE, show_labels = T, print_plot = T)
}
\arguments{
\item{transcript}{transcript in bed format}

\item{exon}{exon coordinates in bed format}

\item{five_UTR}{five_UTR coordinates in bed format}

\item{three_UTR}{three_UTR coordinates in bed format}

\item{mychr}{chromosome}

\item{start_loc}{start location on chromomse}

\item{end_loc}{end location on chromosome}

\item{breaks_wanted}{number of breaks on x-axis}

\item{show_partial_overlap}{include loops that do not entirely overlap the coordinates}

\item{marginvec_mm}{margin around plot, default = c(0,0,0,0)}

\item{y_coord_gene_dist}{distance between genes}

\item{y_top_space}{additional space on top}

\item{y_genename_space}{space between gene name and gene model}

\item{genename_size}{size of gene name}

\item{show_labels}{boolean: show x-axis lables}

\item{print_plot}{boolean: print individual plot}
}
\value{
ggplot object
}
\description{
This function plots the genomic model
}
\examples{
    plot_genome(mychr, ...)


}
\keyword{bedpe}
\keyword{loops,}
