% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeTrackVIz2_main.R
\name{plot_RNA_bedgraph}
\alias{plot_RNA_bedgraph}
\title{RNA bedgraph plotter}
\usage{
plot_RNA_bedgraph(rna_bdg, name_rna_bdg, color_rna_bdg, mychr, start_loc,
  end_loc, binsize = 100, breaks_wanted = 4, y_title = "RPKM",
  chrom_col = "chr", start_col = "start", end_col = "end",
  counts_col = "RPKM", strand_col = NA, min_strand_sign = "-",
  plus_strand_sign = "+", make_minus_strand_negative = T,
  forced_y_min = NULL, forced_y_max = NULL, axis_line_size = 0.2,
  marginvec_mm = c(0, 0, 0, 0), use_barplot = F, alpha_fill = 0.7,
  reverse_strand = FALSE, show_labels = F, print_plot = F)
}
\arguments{
\item{rna_bdg}{bedgraph file for RNA expression data}

\item{name_rna_bdg}{name of sample}

\item{color_rna_bdg}{color of track}

\item{mychr}{chromosome}

\item{start_loc}{start location on chromomse}

\item{end_loc}{end location on chromosome}

\item{binsize}{size of bin}

\item{breaks_wanted}{number of breaks on x-axis}

\item{y_title}{title for y-axis}

\item{chrom_col}{column name for chromosomes}

\item{start_col}{column name for start}

\item{end_col}{column name for end}

\item{counts_col}{column name with counts}

\item{strand_col}{column name with strand info, default = NA}

\item{min_strand_sign}{minus strand sign, default = '-'}

\item{plus_strand_sign}{plus strand sign, default = '+'}

\item{make_minus_strand_negative}{reverse strand}

\item{forced_y_min}{hard minimum y-axis value}

\item{forced_y_max}{hard maximum y-axis value}

\item{axis_line_size}{line size of x-axis}

\item{marginvec_mm}{margin around plot, default = c(0,0,0,0)}

\item{show_labels}{boolean: show x-axis lables}

\item{print_plot}{boolean: print individual plot}
}
\value{
ggplot object
}
\description{
This function plots RNA bedgraph data
}
\examples{
    plot_RNA_bedgraph(rna_bdg)


}
\keyword{RNA,}
\keyword{bedgraph}
