% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeTrackVIz2_main.R
\name{plot_loops}
\alias{plot_loops}
\title{loops or interactions plotter}
\usage{
plot_loops(bedpe, name_bedpe, color_bedpe, mychr, start_loc, end_loc,
  breaks_wanted = 4, y_title = "prob", chrom_col = "chr",
  start_col = "start", end_col = "end", show_partial_overlap = T,
  axis_line_size = 0.2, marginvec_mm = c(0, 0, 0, 0),
  reverse_strand = FALSE, show_labels = F, print_plot = F)
}
\arguments{
\item{bedpe}{bedpe file for DNA loops or interaction data}

\item{name_bedpe}{name of sample}

\item{color_bedpe}{color of track}

\item{mychr}{chromosome}

\item{start_loc}{start location on chromomse}

\item{end_loc}{end location on chromosome}

\item{breaks_wanted}{number of breaks on x-axis}

\item{y_title}{title for y-axis}

\item{chrom_col}{column name for chromosomes}

\item{start_col}{column name for start}

\item{end_col}{column name for end}

\item{show_partial_overlap}{include loops that do not entirely overlap the coordinates}

\item{axis_line_size}{line size of x-axis}

\item{marginvec_mm}{margin around plot, default = c(0,0,0,0)}

\item{show_labels}{boolean: show x-axis lables}

\item{print_plot}{boolean: print individual plot}
}
\value{
ggplot object
}
\description{
This function plots data in bedpe format
}
\examples{
    plot_loops(bedpe)


}
\keyword{bedpe}
\keyword{loops,}
